# Use faster hash map implementation