--- a/LICENSE-THIRD-PARTY.txt
+++ b/LICENSE-THIRD-PARTY.txt
@@ -198,6 +198,7 @@ Library.
 ** lazy_static
 ** rand
 ** itertools
+** hashbrown
                              Apache License
                         Version 2.0, January 2004
                      http://www.apache.org/licenses/