--- a/src/main.rs
+++ b/src/main.rs
@@ -15,14 +15,14 @@ use magic_numbers::*;
 use bloomfilter::*;
 
 use std::time::Instant;
+use pseudoprimes::time::get_elapsed_time;
 
 
 fn main() {
-    let total = Instant::now();
+    let start = Instant::now();
 
     // fp p<=0.001, 32GiB, k=2
     let filter = bloom_t1(&T1_INVERSE);
-    println!("[timing]: bloom_t1 {} milliseconds", total.elapsed().as_millis());
 
     let t2_map = build_t2(filter, &T2);
 
@@ -34,6 +34,5 @@ fn main() {
         println!("Found passing prime {}, vector {:?}", result.pseudoprime, result.factors);
     }
 
-    println!("[total]: Completed in {}ms", total.elapsed().as_millis());
-    println!("Total time: {} seconds, primes found: {}", total.elapsed().as_secs(), results.len());
+    println!("Completed in: {}, primes found: {}", get_elapsed_time(start), results.len());
 }