--- a/src/mulmod.rs
+++ b/src/mulmod.rs
@@ -1,4 +1,4 @@
-// mod.rs Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// bitset.rs Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 #[cfg(all(feature = "unstable", target_arch = "x86_64"))]