*Issue #, if available:*
N/A

*Description of changes:*
While reading [this blog post](https://blog.waffles.space/2018/12/07/deep-dive-into-hashbrown/) about how hash maps and Google's SwissTable work I was curious to see it's impact to this code. Hashbrown is a rust implementation of SwissTable. There is a discussion about using this in the standard rust hash map in https://github.com/rust-lang/rust/issues/55514 but until that get migrated this saves us ~3 seconds. It only affects the final sieve:
Standard:
```
[bloom_t1] Completed 4294967296 in 13811ms
[t2_map] Completed 4294967296 in 6261ms
[final_sieve] Completed in 4272ms
[total]: Completed in 24375ms

[bloom_t1] Completed 4294967296 in 14673ms
[t2_map] Completed 4294967296 in 6162ms
[final_sieve] Completed in 4349ms
[total]: Completed in 25223ms
```
hashbrown:
```
[bloom_t1] Completed 4294967296 in 13715ms
[t2_map] Completed 4294967296 in 6182ms
[final_sieve] Completed in 1898ms
[total]: Completed in 21821ms

[bloom_t1] Completed 4294967296 in 13854ms
[t2_map] Completed 4294967296 in 6156ms
[final_sieve] Completed in 1882ms
[total]: Completed in 21920ms
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.